Move c10:::Device to a standalone header file
Summary: Following the example of refactoring c10::DeviceType, move c10::Device into a standalone header file.
