+++ c10/macros/Macros.h
@@ -94,22 +94,7 @@
   classname(const classname&) = delete;        \
   classname& operator=(const classname&) = delete
 
-#define C10_CONCATENATE_IMPL(s1, s2) s1##s2
-#define C10_CONCATENATE(s1, s2) C10_CONCATENATE_IMPL(s1, s2)
-
 #define C10_MACRO_EXPAND(args) args
-
-/**
- * C10_ANONYMOUS_VARIABLE(str) introduces a new identifier which starts with
- * str and ends with a unique number.
- */
-#ifdef __COUNTER__
-#define C10_UID __COUNTER__
-#define C10_ANONYMOUS_VARIABLE(str) C10_CONCATENATE(str, __COUNTER__)
-#else
-#define C10_UID __LINE__
-#define C10_ANONYMOUS_VARIABLE(str) C10_CONCATENATE(str, __LINE__)
-#endif
 
 #ifdef __has_cpp_attribute
 #define C10_HAS_CPP_ATTRIBUTE(x) __has_cpp_attribute(x)
